Create a 'results' directory
This holds processed data, ready to be make into a graph. It is
intended that processed data is in CSV format, although other
similar formats are ok too. It should be a text (human readable)
format, in case we need to refer back to these numbers at a later
time.

These data files are used by scripts directly to create graphs for
publication.

Signed-off-by: Joel Nider <joel@ece.ubc.ca>