Use PNG format
Updated most of the images to PNG from PDF. PNG is a bitmap format
which is not ideal, but is of high enough resolution for the papers
we publish (800x600 scaled down to about 3 inch width is ~260 DPI).
A vector format is better, but not supported yet by standard tools.
PDF is not a graphics format (it is just a wrapper) and I have no
idea what the actual format is inside the PDF. PNG allows for
previewing in both Overleaf and the OS, whereas PDF does not.
Done while moving scripts from 'asplos21' to 'graphs' and putting
the CSV source files in 'results'.

Signed-off-by: Joel Nider <joel@ece.ubc.ca>